Add support for multiple IdPs
In some cases we might want to authenticate against different IdPs for
diffent resources behind the SP. In order to be able to fetch metadata
from more than one source, we need to turn the variable
mellon_sp_remote_idp_metadata_url from a string to a list.

If a string is provided or a list of one feed then the metadata will be
downloaded as idp.xml ensuring backwards compatibility. If more than one
feeds are in the list then each feed will be downladed as idp_X.xml